8.12. Оптимизируй это (часть 1) Правки по тз проекта
