+++ project/package.json
@@ -9,6 +9,7 @@
     "history": "^5.3.0",
     "http-status-codes": "^2.2.0",
     "leaflet": "^1.9.3",
+    "moment": "^2.29.4",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "react-helmet-async": "^1.3.0",
