+++ project/src/components/property-rating/property-rating.tsx
