+++ project/src/components/rating-place/rating-place.tsx
