+++ project/src/components/review-rating/review-rating.tsx
