+++ project/src/components/top-icons/top-icons.tsx
