+++ project/src/index.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import ReactDOM from 'react-dom/client';
 import {BrowserRouter} from 'react-router-dom';
 import {HelmetProvider} from 'react-helmet-async';
-import {App} from './components/App/App';
+import {App} from './components/app/app';
 import {ToastContainer} from 'react-toastify';
 import { Provider } from 'react-redux';
 import { store } from './store';
