+++ project/src/store/index.ts
@@ -1,9 +1,9 @@
 import {configureStore} from '@reduxjs/toolkit';
 import {NameSpace} from '../consts';
 
-import offersSlice from './slices/OffersSlice/offers.slice';
-import authSlice from './slices/AuthSlice/auth.slice';
-import reviewsSlice from './slices/ReviewsSlice/reviews.slice';
+import offersSlice from './slices/offers-slice/offers.slice';
+import authSlice from './slices/auth-slice/auth.slice';
+import reviewsSlice from './slices/reviews-slice/reviews.slice';
 
 
 export const store = configureStore({
