--- project/src/store/slices/ReviewsSlice/reviews.selectors.ts
@@ -1,4 +1,0 @@
-import { State } from '../../../types/state';
-import { NameSpace } from '../../../consts';
-
-export const getCommentsOnId = ((state: State) => state[NameSpace.Reviews].reviews);
