+++ project/src/store/slices/offers-slice/offers.selectors.ts
